api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "langchain-cli>=0.0.36",
    "langchain-community>=0.3.24",
    "langchain-openai>=0.3.17",
    "langchain-pinecone>=0.2.6",
    "openai>=1.81.0",
    "pymupdf>=1.25.5",
    "python-dotenv>=1.1.0",
    "tiktoken>=0.9.0",
]
